Created interface CharacterService.java changed method's name to more suitable HttpClient changed as private final field and initialized through constructor.
